requirements.txt
asgiref==3.8.1
asttokens==2.4.1
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
gunicorn==23.0.0
h11==0.14.0
ipython==8.29.0
jedi==0.19.1
matplotlib-inline==0.1.7
packaging==24.2
parso==0.8.4
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
tzdata==2024.2
uvicorn==0.32.0
wcwidth==0.2.13
whitenoise==6.8.2
